refactor: move ImageConfigDialog to shared components and update imports in chat-header and image-model-manager for better organization
