+++ surfsense_web/components/settings/image-model-manager.tsx
@@ -3,29 +3,22 @@
 import { useAtomValue } from "jotai";
 import {
 	AlertCircle,
-	Check,
-	ChevronsUpDown,
 	Edit3,
 	Info,
-	Key,
 	Plus,
 	RefreshCw,
 	Trash2,
 	Wand2,
 } from "lucide-react";
-import { useCallback, useMemo, useState } from "react";
-import { toast } from "sonner";
+import { useMemo, useState } from "react";
 import {
-	createImageGenConfigMutationAtom,
 	deleteImageGenConfigMutationAtom,
-	updateImageGenConfigMutationAtom,
 } from "@/atoms/image-gen-config/image-gen-config-mutation.atoms";
 import {
 	globalImageGenConfigsAtom,
 	imageGenConfigsAtom,
 } from "@/atoms/image-gen-config/image-gen-config-query.atoms";
 import { membersAtom, myAccessAtom } from "@/atoms/members/members-query.atoms";
-import { updateLLMPreferencesMutationAtom } from "@/atoms/new-llm-config/new-llm-config-mutation.atoms";
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import {
 	AlertDialog,
@@ -40,43 +33,14 @@ import {
 import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent } from "@/components/ui/card";
-import {
-	Command,
-	CommandEmpty,
-	CommandGroup,
-	CommandInput,
-	CommandItem,
-	CommandList,
-} from "@/components/ui/command";
-import {
-	Dialog,
-	DialogContent,
-	DialogDescription,
-	DialogHeader,
-	DialogTitle,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-} from "@/components/ui/select";
-import { Separator } from "@/components/ui/separator";
 import { Skeleton } from "@/components/ui/skeleton";
 import { Spinner } from "@/components/ui/spinner";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui/tooltip";
-import {
-	getImageGenModelsByProvider,
-	IMAGE_GEN_PROVIDERS,
-} from "@/contracts/enums/image-gen-providers";
 import type { ImageGenerationConfig } from "@/contracts/types/new-llm-config.types";
 import { useMediaQuery } from "@/hooks/use-media-query";
 import { getProviderIcon } from "@/lib/provider-icons";
 import { cn } from "@/lib/utils";
+import { ImageConfigDialog } from "@/components/shared/image-config-dialog";
 
 interface ImageModelManagerProps {
 	searchSpaceId: number;
@@ -92,23 +56,12 @@ function getInitials(name: string): string {
 
 export function ImageModelManager({ searchSpaceId }: ImageModelManagerProps) {
 	const isDesktop = useMediaQuery("(min-width: 768px)");
-	// Image gen config atoms
-	const {
-		mutateAsync: createConfig,
-		isPending: isCreating,
-		error: createError,
-	} = useAtomValue(createImageGenConfigMutationAtom);
-	const {
-		mutateAsync: updateConfig,
-		isPending: isUpdating,
-		error: updateError,
-	} = useAtomValue(updateImageGenConfigMutationAtom);
+
 	const {
 		mutateAsync: deleteConfig,
 		isPending: isDeleting,
 		error: deleteError,
 	} = useAtomValue(deleteImageGenConfigMutationAtom);
-	const { mutateAsync: updatePreferences } = useAtomValue(updateLLMPreferencesMutationAtom);
 
 	const {
 		data: userConfigs,
@@ -119,7 +72,6 @@ export function ImageModelManager({ searchSpaceId }: ImageModelManagerProps) {
 	const { data: globalConfigs = [], isFetching: globalLoading } =
 		useAtomValue(globalImageGenConfigsAtom);
 
-	// Members for user resolution
 	const { data: members } = useAtomValue(membersAtom);
 	const memberMap = useMemo(() => {
 		const map = new Map<string, { name: string; email?: string; avatarUrl?: string }>();
@@ -135,7 +87,6 @@ export function ImageModelManager({ searchSpaceId }: ImageModelManagerProps) {
 		return map;
 	}, [members]);
 
-	// Permissions
 	const { data: access } = useAtomValue(myAccessAtom);
 	const canCreate = useMemo(() => {
 		if (!access) return false;
@@ -147,92 +98,25 @@ export function ImageModelManager({ searchSpaceId }: ImageModelManagerProps) {
 		if (access.is_owner) return true;
 		return access.permissions?.includes("image_generations:delete") ?? false;
 	}, [access]);
-	// Backend uses image_generations:create for update as well
 	const canUpdate = canCreate;
 	const isReadOnly = !canCreate && !canDelete;
 
-	// Local state
 	const [isDialogOpen, setIsDialogOpen] = useState(false);
 	const [editingConfig, setEditingConfig] = useState<ImageGenerationConfig | null>(null);
 	const [configToDelete, setConfigToDelete] = useState<ImageGenerationConfig | null>(null);
 
-	const isSubmitting = isCreating || isUpdating;
 	const isLoading = configsLoading || globalLoading;
-	const errors = [createError, updateError, deleteError, fetchError].filter(Boolean) as Error[];
+	const errors = [deleteError, fetchError].filter(Boolean) as Error[];
 
-	// Form state for create/edit dialog
-	const [formData, setFormData] = useState({
-		name: "",
-		description: "",
-		provider: "",
-		custom_provider: "",
-		model_name: "",
-		api_key: "",
-		api_base: "",
-		api_version: "",
-	});
-	const [modelComboboxOpen, setModelComboboxOpen] = useState(false);
-
-	const resetForm = () => {
-		setFormData({
-			name: "",
-			description: "",
-			provider: "",
-			custom_provider: "",
-			model_name: "",
-			api_key: "",
-			api_base: "",
-			api_version: "",
-		});
+	const openEditDialog = (config: ImageGenerationConfig) => {
+		setEditingConfig(config);
+		setIsDialogOpen(true);
 	};
 
-	const handleFormSubmit = useCallback(async () => {
-		if (!formData.name || !formData.provider || !formData.model_name || !formData.api_key) {
-			toast.error("Please fill in all required fields");
-			return;
-		}
-		try {
-			if (editingConfig) {
-				await updateConfig({
-					id: editingConfig.id,
-					data: {
-						name: formData.name,
-						description: formData.description || undefined,
-						provider: formData.provider as any,
-						custom_provider: formData.custom_provider || undefined,
-						model_name: formData.model_name,
-						api_key: formData.api_key,
-						api_base: formData.api_base || undefined,
-						api_version: formData.api_version || undefined,
-					},
-				});
-			} else {
-				const result = await createConfig({
-					name: formData.name,
-					description: formData.description || undefined,
-					provider: formData.provider as any,
-					custom_provider: formData.custom_provider || undefined,
-					model_name: formData.model_name,
-					api_key: formData.api_key,
-					api_base: formData.api_base || undefined,
-					api_version: formData.api_version || undefined,
-					search_space_id: searchSpaceId,
-				});
-				// Auto-assign newly created config
-				if (result?.id) {
-					await updatePreferences({
-						search_space_id: searchSpaceId,
-						data: { image_generation_config_id: result.id },
-					});
-				}
-			}
-			setIsDialogOpen(false);
-			setEditingConfig(null);
-			resetForm();
-		} catch {
-			// Error handled by mutation
-		}
-	}, [editingConfig, formData, searchSpaceId, createConfig, updateConfig, updatePreferences]);
+	const openNewDialog = () => {
+		setEditingConfig(null);
+		setIsDialogOpen(true);
+	};
 
 	const handleDelete = async () => {
 		if (!configToDelete) return;
@@ -243,30 +127,6 @@ export function ImageModelManager({ searchSpaceId }: ImageModelManagerProps) {
 			// Error handled by mutation
 		}
 	};
-
-	const openEditDialog = (config: ImageGenerationConfig) => {
-		setEditingConfig(config);
-		setFormData({
-			name: config.name,
-			description: config.description || "",
-			provider: config.provider,
-			custom_provider: config.custom_provider || "",
-			model_name: config.model_name,
-			api_key: config.api_key,
-			api_base: config.api_base || "",
-			api_version: config.api_version || "",
-		});
-		setIsDialogOpen(true);
-	};
-
-	const openNewDialog = () => {
-		setEditingConfig(null);
-		resetForm();
-		setIsDialogOpen(true);
-	};
-
-	const selectedProvider = IMAGE_GEN_PROVIDERS.find((p) => p.value === formData.provider);
-	const suggestedModels = getImageGenModelsByProvider(formData.provider);
 
 	return (
 		<div className="space-y-4 md:space-y-6">
@@ -348,31 +208,26 @@ export function ImageModelManager({ searchSpaceId }: ImageModelManagerProps) {
 			{/* Loading Skeleton */}
 			{isLoading && (
 				<div className="space-y-4 md:space-y-6">
-					{/* Your Image Models Section Skeleton */}
 					<div className="space-y-4">
 						<div className="flex items-center justify-between">
 							<Skeleton className="h-6 md:h-7 w-40 md:w-48" />
 							<Skeleton className="h-8 md:h-9 w-32 md:w-36 rounded-md" />
 						</div>
 
-						{/* Cards Grid Skeleton */}
 						<div className="grid gap-3 grid-cols-1 sm:grid-cols-2 xl:grid-cols-3">
 							{["skeleton-a", "skeleton-b", "skeleton-c"].map((key) => (
 								<Card key={key} className="border-border/60">
 									<CardContent className="p-4 flex flex-col gap-3">
-										{/* Header */}
 										<div className="flex items-start justify-between gap-2">
 											<div className="space-y-1.5 flex-1 min-w-0">
 												<Skeleton className="h-4 w-28 md:w-32" />
 												<Skeleton className="h-3 w-40 md:w-48" />
 											</div>
 										</div>
-										{/* Provider + Model */}
 										<div className="flex items-center gap-2">
 											<Skeleton className="h-5 w-16 rounded-full" />
 											<Skeleton className="h-5 w-24 rounded-md" />
 										</div>
-										{/* Footer */}
 										<div className="flex items-center gap-2 pt-2 border-t border-border/40">
 											<Skeleton className="h-3 w-20" />
 											<Skeleton className="h-4 w-4 rounded-full" />
@@ -529,204 +384,18 @@ export function ImageModelManager({ searchSpaceId }: ImageModelManagerProps) {
 				</div>
 			)}
 
-			{/* Create/Edit Dialog */}
-			<Dialog
+			{/* Create/Edit Dialog — shared component */}
+			<ImageConfigDialog
 				open={isDialogOpen}
 				onOpenChange={(open) => {
-					if (!open) {
-						setIsDialogOpen(false);
-						setEditingConfig(null);
-						resetForm();
-					}
+					setIsDialogOpen(open);
+					if (!open) setEditingConfig(null);
 				}}
-			>
-				<DialogContent
-					className="max-w-lg max-h-[90vh] overflow-y-auto"
-					onOpenAutoFocus={(e) => e.preventDefault()}
-				>
-					<DialogHeader>
-						<DialogTitle>{editingConfig ? "Edit Image Model" : "Add Image Model"}</DialogTitle>
-						<DialogDescription>
-							{editingConfig
-								? "Update your image generation model"
-								: "Configure a new image generation model (DALL-E 3, GPT Image 1, etc.)"}
-						</DialogDescription>
-					</DialogHeader>
-
-					<div className="space-y-4 pt-2">
-						{/* Name */}
-						<div className="space-y-2">
-							<Label className="text-sm font-medium">Name *</Label>
-							<Input
-								placeholder="e.g., My DALL-E 3"
-								value={formData.name}
-								onChange={(e) => setFormData((p) => ({ ...p, name: e.target.value }))}
-							/>
-						</div>
-
-						{/* Description */}
-						<div className="space-y-2">
-							<Label className="text-sm font-medium">Description</Label>
-							<Input
-								placeholder="Optional description"
-								value={formData.description}
-								onChange={(e) => setFormData((p) => ({ ...p, description: e.target.value }))}
-							/>
-						</div>
-
-						<Separator />
-
-						{/* Provider */}
-						<div className="space-y-2">
-							<Label className="text-sm font-medium">Provider *</Label>
-							<Select
-								value={formData.provider}
-								onValueChange={(val) =>
-									setFormData((p) => ({ ...p, provider: val, model_name: "" }))
-								}
-							>
-								<SelectTrigger>
-									<SelectValue placeholder="Select a provider" />
-								</SelectTrigger>
-								<SelectContent>
-									{IMAGE_GEN_PROVIDERS.map((p) => (
-										<SelectItem key={p.value} value={p.value}>
-											<div className="flex flex-col">
-												<span className="font-medium">{p.label}</span>
-												<span className="text-xs text-muted-foreground">{p.example}</span>
-											</div>
-										</SelectItem>
-									))}
-								</SelectContent>
-							</Select>
-						</div>
-
-						{/* Model Name */}
-						<div className="space-y-2">
-							<Label className="text-sm font-medium">Model Name *</Label>
-							{suggestedModels.length > 0 ? (
-								<Popover open={modelComboboxOpen} onOpenChange={setModelComboboxOpen}>
-									<PopoverTrigger asChild>
-										<Button
-											variant="outline"
-											role="combobox"
-											className="w-full justify-between font-normal bg-transparent hover:bg-transparent hover:text-foreground"
-										>
-											{formData.model_name || "Select a model"}
-											<ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-										</Button>
-									</PopoverTrigger>
-									<PopoverContent className="w-full p-0" align="start">
-										<Command className="bg-transparent">
-											<CommandInput
-												placeholder="Search a model name"
-												value={formData.model_name}
-												onValueChange={(val) => setFormData((p) => ({ ...p, model_name: val }))}
-											/>
-											<CommandList>
-												<CommandEmpty>
-													<span className="text-xs text-muted-foreground">
-														Type a custom model name
-													</span>
-												</CommandEmpty>
-												<CommandGroup>
-													{suggestedModels.map((m) => (
-														<CommandItem
-															key={m.value}
-															value={m.value}
-															onSelect={() => {
-																setFormData((p) => ({ ...p, model_name: m.value }));
-																setModelComboboxOpen(false);
-															}}
-														>
-															<Check
-																className={cn(
-																	"mr-2 h-4 w-4",
-																	formData.model_name === m.value ? "opacity-100" : "opacity-0"
-																)}
-															/>
-															<span className="font-mono text-sm">{m.value}</span>
-															<span className="ml-2 text-xs text-muted-foreground">{m.label}</span>
-														</CommandItem>
-													))}
-												</CommandGroup>
-											</CommandList>
-										</Command>
-									</PopoverContent>
-								</Popover>
-							) : (
-								<Input
-									placeholder="e.g., dall-e-3"
-									value={formData.model_name}
-									onChange={(e) => setFormData((p) => ({ ...p, model_name: e.target.value }))}
-								/>
-							)}
-						</div>
-
-						{/* API Key */}
-						<div className="space-y-2">
-							<Label className="text-sm font-medium flex items-center gap-1.5">
-								<Key className="h-3.5 w-3.5" /> API Key *
-							</Label>
-							<Input
-								type="password"
-								placeholder="sk-..."
-								value={formData.api_key}
-								onChange={(e) => setFormData((p) => ({ ...p, api_key: e.target.value }))}
-							/>
-						</div>
-
-						{/* API Base (optional) */}
-						<div className="space-y-2">
-							<Label className="text-sm font-medium">API Base URL</Label>
-							<Input
-								placeholder={selectedProvider?.apiBase || "Optional"}
-								value={formData.api_base}
-								onChange={(e) => setFormData((p) => ({ ...p, api_base: e.target.value }))}
-							/>
-						</div>
-
-						{/* API Version (Azure) */}
-						{formData.provider === "AZURE_OPENAI" && (
-							<div className="space-y-2">
-								<Label className="text-sm font-medium">API Version (Azure)</Label>
-								<Input
-									placeholder="2024-02-15-preview"
-									value={formData.api_version}
-									onChange={(e) => setFormData((p) => ({ ...p, api_version: e.target.value }))}
-								/>
-							</div>
-						)}
-
-						{/* Actions */}
-						<div className="flex justify-end gap-3 pt-4 border-t">
-							<Button
-								variant="secondary"
-								onClick={() => {
-									setIsDialogOpen(false);
-									setEditingConfig(null);
-									resetForm();
-								}}
-							>
-								Cancel
-							</Button>
-							<Button
-								onClick={handleFormSubmit}
-								disabled={
-									isSubmitting ||
-									!formData.name ||
-									!formData.provider ||
-									!formData.model_name ||
-									!formData.api_key
-								}
-							>
-								{isSubmitting ? <Spinner size="sm" className="mr-2" /> : null}
-								{editingConfig ? "Save Changes" : "Create & Use"}
-							</Button>
-						</div>
-					</div>
-				</DialogContent>
-			</Dialog>
+				config={editingConfig}
+				isGlobal={false}
+				searchSpaceId={searchSpaceId}
+				mode={editingConfig ? "edit" : "create"}
+			/>
 
 			{/* Delete Confirmation */}
 			<AlertDialog
